[level unrated] Title: 두 수의 나눗셈, Time: 0.18 ms, Memory: 31.4 MB -BaekjoonHub

+++ 프로그래머스/unrated/120806. 두 수의 나눗셈/두 수의 나눗셈.cs
@@ -5,7 +5,7 @@ public class Solution {
         int answer = 0;
         float result;
         
-        result = ((float)num1 / (float)num2) * 1000;
+        result = ((float)num1 / num2) * 1000;
         answer = (int)result;
 
         return answer;
